リファクタリング Page componentからdragItemのデータをバケツリレーする
