+++ work/resources/ts/containers/organisms/DragBoardList.tsx
@@ -1,71 +1,31 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { DragDropContext, Droppable } from 'react-beautiful-dnd';
 import { DragBoardItem } from '../../components/molecules/DragBoardItem';
 import { AddIcon } from '../../components/atoms/AddIcon';
 import { ModalWindow } from '../../components/molecules/ModalWindow';
 
-type ItemType = {
-  id: string;
-  content: string;
-}
 type BoardListProps = {
   items?: any,
   onClick: (event: any) => void,
   onClose: (event: any) => void,
   modalOpen: boolean,
+  onDragEnd: any,
 }
-
-const reorder = (
-  list: ItemType[],
-  startIndex: number,
-  endIndex: number
-): ItemType[] => {
-  const result = Array.from(list);
-  const [removed] = result.splice(startIndex, 1);
-  result.splice(endIndex, 0, removed);
-
-  return result;
-};
 
 export const DragBoardList = React.memo<BoardListProps> (({
   items,
   onClick,
   onClose,
-  modalOpen
+  modalOpen,
+  onDragEnd,
 }) => {
-  const initial: ItemType[] = Array.from({ length: 10 }, (v, k) => k).map(k => {
-    return {
-      id: `id-${k}`,
-      content: `Item ${k}`
-    }
-  })
-  const [state, setState] = useState({ items: initial });
-
-  const onDragEnd = (result: any) => {
-    if (!result.destination) {
-      return;
-    }
-
-    if (result.destination.index === result.source.index) {
-      return;
-    }
-
-    const items = reorder(
-      state.items,
-      result.source.index,
-      result.destination.index
-    );
-
-    setState({ items });
-  };
-
   return (
     <>
       <DragDropContext onDragEnd={onDragEnd}>
         <Droppable droppableId="list">
           {provided => (
             <div ref={provided.innerRef}>
-              {state.items.map((
+              {items.map((
                 item: {
                   id: string,
                   content: string,
