+++ work/resources/ts/containers/pages/HomePage.tsx
@@ -8,7 +8,24 @@ import {
   Grid,
 } from '@material-ui/core';
 
-export const HomePage: React.FC = () => {
+type ItemType = {
+  id: string;
+  content: string;
+}
+
+const reorder = (
+  list: ItemType[],
+  startIndex: number,
+  endIndex: number
+): ItemType[] => {
+  const result = Array.from(list);
+  const [removed] = result.splice(startIndex, 1);
+  result.splice(endIndex, 0, removed);
+
+  return result;
+};
+
+export const HomePage = React.memo (() => {
   // classNameのインポート
   const { useStyles } = useContext<any>(StylesContext);
   const classes = useStyles();
@@ -22,6 +39,33 @@ export const HomePage: React.FC = () => {
     setModalOpen(false);
   };
 
+  // dragitemのデータ
+  const initial: ItemType[] = Array.from({ length: 5 }, (v, k) => k).map(k => {
+    return {
+      id: `id-${k}`,
+      content: `Item ${k}`
+    }
+  })
+  const [state, setState] = useState({ items: initial });
+
+  const onDragEnd = (result: any) => {
+    if (!result.destination) {
+      return;
+    }
+
+    if (result.destination.index === result.source.index) {
+      return;
+    }
+
+    const items = reorder(
+      state.items,
+      result.source.index,
+      result.destination.index
+    );
+
+    setState({ items });
+  };
+
   return (
     <>
       <CssBaseline />
@@ -33,6 +77,8 @@ export const HomePage: React.FC = () => {
           <Grid item xs={6}>
             <Container maxWidth='xl'>
               <DragBoardList
+                items={state.items}
+                onDragEnd={onDragEnd}
                 onClick={() => {
                   modalOpen();
                 }}
@@ -52,4 +98,4 @@ export const HomePage: React.FC = () => {
       </Container>
     </>
   )
-}
+})
